+++ media/smart-card/src/utils/__tests__/analytics.test.ts
@@ -3,6 +3,7 @@ import {
   ANALYTICS_CHANNEL,
   resolvedEvent,
   unresolvedEvent,
+  instrumentEvent,
 } from '../analytics';
 import { AnalyticsPayload } from '../types';
 import { APIError } from '../../client/errors';
@@ -92,6 +93,66 @@ describe('unresolvedEvent', () => {
           "packageName": "@atlaskit/smart-card",
           "packageVersion": "999.9.9",
           "reason": "fail",
+          "resourceType": "file-type",
+        },
+        "eventType": "operational",
+      }
+    `);
+  });
+});
+
+describe('instrumentEvent', () => {
+  test('does not return ResolveUnsupportedError unresolved events', () => {
+    const error = new APIError(
+      'error',
+      'localhost',
+      'something wrong',
+      'ResolveUnsupportedError',
+    );
+    const event = instrumentEvent(
+      'id',
+      'errored',
+      'definition-id',
+      'extension-key',
+      'file-type',
+      error,
+    );
+    expect(event).toBeUndefined();
+  });
+
+  test('returns unresolved events for events apart from ResolveUnsupportedError unresolved events', () => {
+    const error = new APIError(
+      'error',
+      'localhost',
+      'something wrong',
+      'ResolveFailedError',
+    );
+    const event = instrumentEvent(
+      'id',
+      'errored',
+      'definition-id',
+      'extension-key',
+      'file-type',
+      error,
+    );
+    expect(event).toMatchInlineSnapshot(`
+      Object {
+        "action": "unresolved",
+        "actionSubject": "smartLink",
+        "attributes": Object {
+          "componentName": "smart-cards",
+          "definitionId": "definition-id",
+          "duration": undefined,
+          "error": Object {
+            "kind": "error",
+            "message": "something wrong",
+            "type": "ResolveFailedError",
+          },
+          "extensionKey": "extension-key",
+          "id": "id",
+          "packageName": "@atlaskit/smart-card",
+          "packageVersion": "999.9.9",
+          "reason": "errored",
           "resourceType": "file-type",
         },
         "eventType": "operational",