+++ editor/editor-mobile-bridge/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/editor-mobile-bridge
 
+## 24.0.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 24.0.2
 
 ### Patch Changes